PR #3204: Migrate MaxText/train_compile.py to maxtext/trainers/pre_train/train_compile.py
Imported from GitHub PR https://github.com/AI-Hypercomputer/maxtext/pull/3204

# Description

* Move `MaxText/train_compile.py` to `maxtext/trainers/pre_train/train_compile.py`
* Create shim in MaxText/train.py to support old command. Include a deprecation warning
    * TODO: Add deprecation dates to this and other shims we are adding
* Next steps: Update [existing `EXECUTABLE` commands](https://github.com/AI-Hypercomputer/maxtext/blob/main/src/maxtext/configs/tpu/v4/22b.sh#L58) to call new path for train/train_compile. Leaving this off for now to keep PR smaller. The old commands still work for these

NOTE: We will fix the ungrouped import lint errors at the end of restructuring. Otherwise we will just be swapping the order of imports repeatedly

# Tests

Both new and old commands working as expected.

New command:
```
python3 -m maxtext.trainers.pre_train.train_compile src/maxtext/configs/base.yml \
    compile_topology=v5e-256 \
    compile_topology_num_slices=2 \
    global_parameter_scale=16 \
    per_device_batch_size=4
```

Old command:
```
python3 -m MaxText.train_compile src/maxtext/configs/base.yml \
    compile_topology=v5e-256 \
    compile_topology_num_slices=2 \
    global_parameter_scale=16 \
    per_device_batch_size=4
```

# Checklist

Before submitting this PR, please make sure (put X in square brackets):
- [x] I have performed a self-review of my code. For an optional AI review, add the `gemini-review` label.
- [x] I have necessary comments in my code, particularly in hard-to-understand areas.
- [x] I have run end-to-end tests tests and provided workload links above if applicable.
- [x] I have made or will make corresponding changes to the doc if needed, including adding new documentation pages to the relevant Table of Contents (toctree directive) as explained in [our documentation](https://maxtext.readthedocs.io/en/latest/development.html#adding-new-documentation-files).

Copybara import of the project:

--
9dc270447a04df72b898edf5fc5494eece702897 by Branden Vandermoon <bvandermoon@google.com>:

Migrate MaxText/train_compile.py to maxtext/trainers/pre_train/train_compile.py

Merging this change closes #3204

FUTURE_COPYBARA_INTEGRATE_REVIEW=https://github.com/AI-Hypercomputer/maxtext/pull/3204 from AI-Hypercomputer:bvandermoon-repo-restructure 9dc270447a04df72b898edf5fc5494eece702897
PiperOrigin-RevId: 873125760
