+++ tests/integration/aot_identical_test.py
@@ -27,7 +27,7 @@ import hashlib
 import re
 import jax
 from tests.utils.test_helpers import get_test_config_path
-from MaxText import train_compile
+from maxtext.trainers.pre_train import train_compile
 from maxtext.trainers.pre_train import train
