+++ tests/integration/xaot_test.py
@@ -25,7 +25,7 @@ import os
 import shutil
 import jax
 from tests.utils.test_helpers import get_test_config_path
-from MaxText import train_compile
+from maxtext.trainers.pre_train import train_compile
 from maxtext.trainers.pre_train import train
