+++ tests/unit/diloco_test.py
@@ -30,7 +30,7 @@ import optax
 import pytest
 
 from MaxText.pyconfig import initialize_pydantic
-from MaxText.train_compile import main as train_compile_main
+from maxtext.trainers.pre_train.train_compile import main as train_compile_main
 from maxtext.trainers.diloco import diloco
 from tests.utils.test_helpers import get_test_config_path
