+++ tests/unit/sharding_compare_test.py
@@ -27,7 +27,7 @@ from MaxText import pyconfig
 from MaxText.globals import MAXTEXT_PKG_DIR
 from maxtext.layers import quantizations
 from maxtext.models import models
-from MaxText.train_compile import get_shaped_inputs, get_topology_mesh, validate_config
+from maxtext.trainers.pre_train.train_compile import get_shaped_inputs, get_topology_mesh, validate_config
 from tests.utils.sharding_dump import TEST_CASES, load_json, named_shardings_to_json, partition_specs_to_json
 import pytest
