+++ tests/unit/train_compile_test.py
@@ -25,7 +25,7 @@ from tempfile import gettempdir
 
 import pytest
 
-from MaxText.train_compile import main as train_compile_main
+from maxtext.trainers.pre_train.train_compile import main as train_compile_main
 from tests.utils.test_helpers import get_test_config_path
 
 pytestmark = [pytest.mark.external_training]
